refactor: 트랜잭션과 비즈니스 로직 클래스로 분리
